+++ src/main/java/de/maxhenkel/audioplayer/config/ServerConfig.java
@@ -8,17 +8,21 @@ public class ServerConfig {
     public final ConfigEntry<String> filebinUrl;
     public final ConfigEntry<Integer> maxUploadSize;
     public final ConfigEntry<Integer> uploadPermissionLevel;
-    public final ConfigEntry<Integer> musicDiscPermissionLevel;
+    public final ConfigEntry<Integer> applyToItemPermissionLevel;
     public final ConfigEntry<Boolean> jukeboxHopperInteraction;
     public final ConfigEntry<Boolean> jukeboxDispenserInteraction;
+    public final ConfigEntry<Integer> goatHornCooldown;
+    public final ConfigEntry<Double> goatHornRange;
 
     public ServerConfig(ConfigBuilder builder) {
         filebinUrl = builder.stringEntry("filebin_url", "https://filebin.net/");
         maxUploadSize = builder.integerEntry("max_upload_size", 1000 * 1000 * 50, 1, Integer.MAX_VALUE);
         uploadPermissionLevel = builder.integerEntry("upload_permission_level", 0, 0, Integer.MAX_VALUE);
-        musicDiscPermissionLevel = builder.integerEntry("music_disc_permission_level", 0, 0, Integer.MAX_VALUE);
+        applyToItemPermissionLevel = builder.integerEntry("apply_to_item_permission_level", 0, 0, Integer.MAX_VALUE);
         jukeboxHopperInteraction = builder.booleanEntry("jukebox_hopper_interaction", true);
         jukeboxDispenserInteraction = builder.booleanEntry("jukebox_dispenser_interaction", true);
+        goatHornCooldown = builder.integerEntry("goat_horn_cooldown", 140, 1, Short.MAX_VALUE);
+        goatHornRange = builder.doubleEntry("goat_horn_range", 256F, 1F, Integer.MAX_VALUE);
     }
 
 }